perf_FIX-lessRAM
- Foram diminuidos valores e Ajustados constantes para ocupar menos espaço na memoria do pic.
- Uso do static para as constantes da função, ocupa menos memoria.

+++ sync.h
@@ -7,7 +7,7 @@
 // Padrão POSIX
 
 typedef struct sem {
-    int contador;
+    uint8_t contador;
     uint8_t fila[MAX_USER_TASKS];
     uint8_t pos_input;
     uint8_t pos_output;
